Fix visual studio warnings

+++ src/Memory.cpp
@@ -211,9 +211,9 @@ bool Memory::IsBiosLoaded()
 
 void Memory::CheckBreakpoints(u16 address, bool write)
 {
-    long unsigned int size = m_BreakpointsMem.size();
+    size_t size = m_BreakpointsMem.size();
 
-    for (long unsigned int b = 0; b < size; b++)
+    for (size_t b = 0; b < size; b++)
     {
         if (write && !m_BreakpointsMem[b].write)
             continue;
