+++ src/Processor.cpp
@@ -465,9 +465,9 @@ bool Processor::Disassemble(u16 address)
     }
     else
     {
-        long unsigned int size = breakpoints->size();
+        size_t size = breakpoints->size();
 
-        for (long unsigned int b = 0; b < size; b++)
+        for (size_t b = 0; b < size; b++)
         {
             if ((*breakpoints)[b] == record)
             {
